fix: filter locales
