+++ src/runtime/server/storage.ts
@@ -199,6 +199,15 @@ export const createServerQueryFetch = <T = ParsedContent>(event: H3Event, path?:
     query.sort({ _file: 1, $numeric: true })
   }
 
+  // Filter by locale if:
+  // - locales are defined
+  // - query doesn't already have a locale filter
+  if (contentConfig.locales.length) {
+    const queryLocale = query.params().where?.find(w => w._locale)?._locale
+    const locale = queryLocale || contentConfig.defaultLocale
+    query.locale(locale)
+  }
+
   return createPipelineFetcher<T>(() => getIndexedContentsList<T>(event, query))(query)
 }
