+++ test/custom-api-base.test.ts
@@ -1,7 +1,6 @@
 import { fileURLToPath } from 'url'
-import { assert, test, describe, expect, vi } from 'vitest'
+import { test, describe, expect, vi } from 'vitest'
 import { setup, $fetch } from '@nuxt/test-utils'
-import { hash } from 'ohash'
 import { testMarkdownParser } from './features/parser-markdown'
 import { testPathMetaTransformer } from './features/transformer-path-meta'
 import { testYamlParser } from './features/parser-yaml'
@@ -18,6 +17,7 @@ import { testHighlighter } from './features/highlighter'
 import { testMarkdownRenderer } from './features/renderer-markdown'
 import { testParserOptions } from './features/parser-options'
 import { testComponents } from './features/components'
+import { testLocales } from './features/locales'
 
 const spyConsoleWarn = vi.spyOn(global.console, 'warn')
 const apiBaseURL = '/my-content-api'
@@ -36,92 +36,9 @@ describe('Custom api baseURL', async () => {
     }
   })
 
-  const QUERY_ENDPOINT = `${apiBaseURL}/query`
-  const fetchDocument = (_id: string) => {
-    const params = { first: true, where: { _id } }
-    const qid = hash(params)
-    return $fetch(`${QUERY_ENDPOINT}/${qid}`, {
-      params: { _params: JSON.stringify(params) }
-    })
-  }
-
-  test('List contents', async () => {
-    const params = { only: '_id' }
-    const qid = hash(params)
-    const docs = await $fetch(`${QUERY_ENDPOINT}/${qid}`, {
-      params: { _params: JSON.stringify(params) }
-    })
-
-    const ids = docs.map(doc => doc._id)
-
-    assert(ids.length > 0)
-    assert(ids.includes('content:index.md'))
-
-    // Ignored files should be listed
-    assert(ids.includes('content:.dot-ignored.md') === false, 'Ignored files with `.` should not be listed')
-    assert(ids.includes('content:-dash-ignored.md') === false, 'Ignored files with `-` should not be listed')
-
-    assert(ids.includes('fa-ir:fa:hello.md') === true, 'Files with `fa-ir` prefix should be listed')
-  })
-
-  test('Get contents index', async () => {
-    const index = await fetchDocument('content:index.md')
-
-    expect(index).toHaveProperty('body')
-
-    expect(index.body).toMatchSnapshot('basic-index-body')
-  })
-
-  test('Get ignored contents', async () => {
-    const ignored = await fetchDocument('content:.dot-ignored.md').catch(_err => null)
-
-    expect(ignored).toBeNull()
-  })
-
-  test('Search contents using `locale` helper', async () => {
-    const fa = await $fetch('/locale-fa')
-
-    expect(fa).toContain('fa-ir:fa:hello.md')
-    expect(fa).not.toContain('content:index.md')
-
-    const en = await $fetch('/locale-en')
-
-    expect(en).not.toContain('fa-ir:fa:hello.md')
-    expect(en).toContain('content:index.md')
-  })
-
-  test('Use default locale for unscoped contents', async () => {
-    const index = await fetchDocument('content:index.md')
-
-    expect(index).toMatchObject({
-      _locale: 'en'
-    })
-  })
-
   test('Multi part path', async () => {
     const html = await $fetch('/features/multi-part-path')
     expect(html).contains('Persian')
-  })
-
-  test('Empty slot', async () => {
-    const html = await $fetch('/features/empty-slot')
-    expect(html).contains('Nullish Document!!!')
-    expect(html).contains('Empty Child!!!')
-  })
-
-  test('<ContentDoc> head management (if same path)', async () => {
-    const html = await $fetch('/head')
-    expect(html).contains('<title>Head overwritten</title>')
-    expect(html).contains('<meta property="og:image" content="https://picsum.photos/200/300">')
-    expect(html).contains('<meta name="description" content="Description overwritten">')
-    expect(html).contains('<meta property="og:image" content="https://picsum.photos/200/300">')
-  })
-
-  test('<ContentDoc> head management (not same path)', async () => {
-    const html = await $fetch('/bypass-head')
-    expect(html).not.contains('<title>Head overwritten</title>')
-    expect(html).not.contains('<meta property="og:image" content="https://picsum.photos/200/300">')
-    expect(html).not.contains('<meta name="description" content="Description overwritten"><meta property="og:image" content="https://picsum.photos/200/300">')
   })
 
   test('Partials specials chars', async () => {
@@ -139,6 +56,8 @@ describe('Custom api baseURL', async () => {
     expect(spyConsoleWarn).toHaveBeenCalled()
     expect(spyConsoleWarn).toHaveBeenCalledWith('Ignoring [content:with-\'invalid\'-char.md]. File name should not contain any of the following characters: \', ", ?, #, /')
   })
+
+  testLocales()
 
   testComponents()
