+++ test/features/renderer-markdown.ts
@@ -67,5 +67,26 @@ export const testMarkdownRenderer = () => {
       const html = await $fetch('/features/slotted-content-renderer')
       expect(html).contains('The default slot is overridden')
     })
+
+    test('Empty slot', async () => {
+      const html = await $fetch('/features/empty-slot')
+      expect(html).contains('Nullish Document!!!')
+      expect(html).contains('Empty Child!!!')
+    })
+
+    test('<ContentDoc> head management (if same path)', async () => {
+      const html = await $fetch('/head')
+      expect(html).contains('<title>Head overwritten</title>')
+      expect(html).contains('<meta property="og:image" content="https://picsum.photos/200/300">')
+      expect(html).contains('<meta name="description" content="Description overwritten">')
+      expect(html).contains('<meta property="og:image" content="https://picsum.photos/200/300">')
+    })
+
+    test('<ContentDoc> head management (not same path)', async () => {
+      const html = await $fetch('/bypass-head')
+      expect(html).not.contains('<title>Head overwritten</title>')
+      expect(html).not.contains('<meta property="og:image" content="https://picsum.photos/200/300">')
+      expect(html).not.contains('<meta name="description" content="Description overwritten"><meta property="og:image" content="https://picsum.photos/200/300">')
+    })
   })
 }
